fix eprintf warnings
aligned_u64 was hard-coded "unsigned long long" in tgtif.c this
breaks compilation on 64bit ARCHES (eg. x86_64). define it to work
well on all platforms, just like the Kernel does it.
[This problem is caused by another problem:
  Including linux/type.h before sys/types.h will not compile, only
  the reverse order will. But when doing so lots of linux/type.h
  definitions are ifdefed out, including the definition of
  aligned_u64. This is why it as to be redefined by tgtif.c.]

And another place that needed an "%zu"

+++ usr/tgtif.c
@@ -34,7 +34,7 @@
 #include <sys/sysmacros.h>
 #include <linux/types.h>
 #ifndef aligned_u64
-#define aligned_u64 unsigned long long __attribute__((aligned(8)))
+#define aligned_u64 uint64_t __attribute__((aligned(8)))
 #endif
 #include <scsi/scsi_tgt_if.h>
